+++ db/user.go
@@ -12,8 +12,8 @@ type User struct {
 	UserName       string  `gorm:"index:idx_username,unique;type:varchar(40);not null" json:"username"`
 	Password       string  `gorm:"type:varchar(256);not null" json:"password"`
 	FavoriteVideos []Video `gorm:"many2many:user_favorite_videos" json:"favorite_videos"`
-	FollowCount    int     `gorm:"default:0" json:"follow_count"`
-	FollowerCount  int     `gorm:"default:0" json:"follower_count"`
+	FollowCount    int64   `gorm:"default:0" json:"follow_count"`
+	FollowerCount  int64   `gorm:"default:0" json:"follower_count"`
 }
 
 func (User) TableName() string {
